+++ src/Application/Contracts/Bitrix24Accounts/Entity/Bitrix24AccountInterface.php
@@ -57,7 +57,7 @@ interface Bitrix24AccountInterface
      */
     public function getAuthToken(): AuthToken;
 
-    
+
     public function renewAuthToken(RenewedAuthToken $renewedAuthToken): void;
 
     /**
@@ -76,16 +76,31 @@ interface Bitrix24AccountInterface
     public function changeDomainUrl(string $newDomainUrl): void;
 
     /**
-     * @param non-empty-string $applicationToken Application installed on portal and finish installation flow,  set status «active»
+     * Finish application installation
+     *
+     * Application installed on portal and finish installation flow, method must set status to «active»
+     * If You installed application without UI, You already have application token in OnApplicationInstall event
+     * If You installed application with UI, You can finish installation flow without a token and receive it in a separate request with OnApplicationInstall event
+     *
+     * @param non-empty-string|null $applicationToken
      * @throws InvalidArgumentException
      */
-    public function applicationInstalled(string $applicationToken): void;
+    public function applicationInstalled(?string $applicationToken): void;
 
     /**
-     * @param string $applicationToken Application uninstalled from portal, set status «deleted»
+     * Set application token from OnApplicationInstall event
+     *
+     * You must before finish installation with call applicationInstalled method
+     *
+     * @param non-empty-string $applicationToken
+     */
+    public function setApplicationToken(string $applicationToken): void;
+
+    /**
+     * @param string|null $applicationToken Application uninstalled from portal, set status «deleted»
      * @throws InvalidArgumentException
      */
-    public function applicationUninstalled(string $applicationToken): void;
+    public function applicationUninstalled(?string $applicationToken): void;
 
     /**
      * Check is application token valid